+++ bytestring/CHANGES.md
@@ -4,6 +4,7 @@
 
 - Minimum supported Rust version (MSRV) is now 1.88.
 - Improve `From<ByteString> for String` performance.
+- Improve ASCII `TryFrom` conversions performance.
 
 ## 1.5.0